Internalize unnecessarily exported macros
